[ deploy ] : fix error && deploy code 🚀

+++ lib/actions/user.action.ts
@@ -1,72 +1,15 @@
-"use server";
+"use server"
 
+import { FilterQuery } from "mongoose";
 import User from "@/database/user.model";
-import { connectToDatabase } from "../mongoose";
-import {
-  CreateUserParams,
-  DeleteUserParams,
-  GetAllUsersParams,
-  GetSavedQuestionsParams,
-  GetUserByIdParams,
-  GetUserStatsParams,
-  ToggleSaveQuestionParams,
-  UpdateUserParams,
-} from "./shared";
+import { connectToDatabase } from "../mongoose"
+import { CreateUserParams, DeleteUserParams, GetAllUsersParams, GetSavedQuestionsParams, GetUserByIdParams, GetUserStatsParams, ToggleSaveQuestionParams, UpdateUserParams } from "./shared.types";
 import { revalidatePath } from "next/cache";
 import Question from "@/database/question.model";
 import Tag from "@/database/tag.model";
-import { FilterQuery } from "mongoose";
 import Answer from "@/database/answer.model";
 import { BadgeCriteriaType } from "@/types";
 import { assignBadges } from "../utils";
-
-export async function getAllUsers(params: GetAllUsersParams) {
-  try {
-    connectToDatabase();
-
-    const { searchQuery, filter, page = 1, pageSize = 2 } = params;
-    const skipAmount = (page - 1) * pageSize;
-
-    const query: FilterQuery<typeof User> = {};
-
-    if (searchQuery) {
-      query.$or = [
-        { name: { $regex: new RegExp(searchQuery, "i") } },
-        { username: { $regex: new RegExp(searchQuery, "i") } },
-      ];
-    }
-
-    let sortOptions = {};
-
-    switch (filter) {
-      case "new_users":
-        sortOptions = { createdAt: -1 };
-        break;
-      case "old_users":
-        sortOptions = { createdAt: 1 };
-        break;
-      case "top_contributors":
-        sortOptions = { reputation: -1 };
-        break;
-
-      default:
-        break;
-    }
-
-    const users = await User.find(query)
-      .sort(sortOptions)
-      .skip(skipAmount)
-      .limit(pageSize);
-
-    const totalUsers = await User.countDocuments(query);
-    const isNext = totalUsers > skipAmount + users.length;
-
-    return { users, isNext };
-  } catch (error) {
-    console.log(error);
-    throw error;
-  }
-}
 
 export async function getUserById(params: any) {
   try {
@@ -88,6 +31,7 @@ export async function createUser(userData: CreateUserParams) {
     connectToDatabase();
 
     const newUser = await User.create(userData);
+
     return newUser;
   } catch (error) {
     console.log(error);
@@ -100,6 +44,7 @@ export async function updateUser(params: UpdateUserParams) {
     connectToDatabase();
 
     const { clerkId, updateData, path } = params;
+
     await User.findOneAndUpdate({ clerkId }, updateData, {
       new: true,
     });
@@ -114,25 +59,77 @@ export async function updateUser(params: UpdateUserParams) {
 export async function deleteUser(params: DeleteUserParams) {
   try {
     connectToDatabase();
+
     const { clerkId } = params;
 
     const user = await User.findOneAndDelete({ clerkId });
 
-    if (!user) {
-      throw new Error("User not found");
+    if(!user) {
+      throw new Error('User not found');
     }
 
-    const userQuestionIds = await Question.find({
-      author: user._id,
-    }).distinct("_id");
+    // Delete user from database
+    // and questions, answers, comments, etc.
 
-    await Question.deleteMany({
-      author: user._id,
-    });
+    // get user question ids
+    // const userQuestionIds = await Question.find({ author: user._id}).distinct('_id');
+
+    // delete user questions
+    await Question.deleteMany({ author: user._id });
+
+    // TODO: delete user answers, comments, etc.
 
     const deletedUser = await User.findByIdAndDelete(user._id);
 
     return deletedUser;
+  } catch (error) {
+    console.log(error);
+    throw error;
+  }
+}
+
+export async function getAllUsers(params: GetAllUsersParams) {
+  try {
+    connectToDatabase();
+
+    const { searchQuery, filter, page = 1, pageSize = 10 } = params;
+    const skipAmount = (page - 1) * pageSize;
+
+    const query: FilterQuery<typeof User> = {};
+
+    if(searchQuery) {
+      query.$or = [
+        { name: { $regex: new RegExp(searchQuery, 'i') }},
+        { username: { $regex: new RegExp(searchQuery, 'i') }},
+      ]
+    }
+
+    let sortOptions = {};
+
+    switch (filter) {
+      case "new_users":
+        sortOptions = { joinedAt: -1 }
+        break;
+      case "old_users":
+        sortOptions = { joinedAt: 1 }
+        break;
+      case "top_contributors":
+        sortOptions = { reputation: -1 }
+        break;
+    
+      default:
+        break;
+    }
+
+    const users = await User.find(query)
+      .sort(sortOptions)
+      .skip(skipAmount)
+      .limit(pageSize)
+
+    const totalUsers = await User.countDocuments(query);
+    const isNext = totalUsers > skipAmount + users.length;
+
+    return { users, isNext };
   } catch (error) {
     console.log(error);
     throw error;
@@ -147,29 +144,27 @@ export async function toggleSaveQuestion(params: ToggleSaveQuestionParams) {
 
     const user = await User.findById(userId);
 
-    if (!user) {
-      throw new Error("User not found");
+    if(!user) {
+      throw new Error('User not found');
     }
 
-    const isquestionSaved = user.saved.includes(questionId);
+    const isQuestionSaved = user.saved.includes(questionId);
 
-    if (isquestionSaved) {
-      await User.findByIdAndUpdate(
-        userId,
-        {
-          $pull: { saved: questionId },
-        },
+    if(isQuestionSaved) {
+      // remove question from saved
+      await User.findByIdAndUpdate(userId, 
+        { $pull: { saved: questionId }},
         { new: true }
-      );
+      )
     } else {
-      await User.findByIdAndUpdate(
-        userId,
-        { $addToSet: { saved: questionId } },
+      // add question to saved
+      await User.findByIdAndUpdate(userId, 
+        { $addToSet: { saved: questionId }},
         { new: true }
-      );
+      )
     }
 
-    revalidatePath(path);
+    revalidatePath(path)
   } catch (error) {
     console.log(error);
     throw error;
@@ -180,42 +175,41 @@ export async function getSavedQuestions(params: GetSavedQuestionsParams) {
   try {
     connectToDatabase();
 
-    const { clerkId, searchQuery, filter, page = 1, pageSize = 3 } = params;
+    const { clerkId, searchQuery, filter, page = 1, pageSize = 20 } = params;
+
     const skipAmount = (page - 1) * pageSize;
-
+    
     const query: FilterQuery<typeof Question> = searchQuery
-      ? { title: { $regex: new RegExp(searchQuery, "i") } }
-      : {};
+      ? { title: { $regex: new RegExp(searchQuery, 'i') } }
+      : { };
 
-    let sortOptions = {};
+      let sortOptions = {};
 
-    switch (filter) {
-      case "most_recent":
-        sortOptions = { createdAt: -1 };
-        break;
+      switch (filter) {
+        case "most_recent":
+          sortOptions = { createdAt: -1 }
+          break;
+        case "oldest":
+          sortOptions = { createdAt: 1 }
+          break;
+        case "most_voted":
+          sortOptions = { upvotes: -1 }
+          break;
+        case "most_viewed":
+          sortOptions = { views: -1 }
+          break;
+        case "most_answered":
+          sortOptions = { answers: -1 }
+          break;
+      
+        default:
+          break;
+      }
 
-      case "oldest":
-        sortOptions = { createdAt: 1 };
-        break;
-
-      case "most_voted":
-        sortOptions = { upvotes: -1 };
-        break;
-
-      case "most_viewed":
-        sortOptions = { views: -1 };
-        break;
-
-      case "most_answered":
-        sortOptions = { answers: -1 };
-        break;
-
-      default:
-        break;
-    }
-
-    const user = await User.findOne({ clerkId }).populate({
-      path: "saved",
+    const user = await User
+    .findOne({ clerkId })
+    .populate({
+      path: 'saved',
       match: query,
       options: {
         sort: sortOptions,
@@ -223,15 +217,15 @@ export async function getSavedQuestions(params: GetSavedQuestionsParams) {
         limit: pageSize + 1,
       },
       populate: [
-        { path: "tags", model: Tag, select: "_id name" },
-        { path: "author", model: User, select: "_id clerkId name picture" },
-      ],
-    });
+        { path: 'tags', model: Tag, select: "_id name" },
+        { path: 'author', model: User, select: '_id clerkId name picture'}
+      ]
+    })
 
     const isNext = user.saved.length > pageSize;
-
-    if (!user) {
-      throw new Error("User not found");
+    
+    if(!user) {
+      throw new Error('User not found');
     }
 
     const savedQuestions = user.saved;
@@ -251,72 +245,50 @@ export async function getUserInfo(params: GetUserByIdParams) {
 
     const user = await User.findOne({ clerkId: userId });
 
-    if (!user) {
-      throw new Error("User not found");
+    if(!user) {
+      throw new Error('User not found');
     }
 
-    const totalQuestions = await Question.countDocuments({ author: user._id });
-
+    const totalQuestions = await Question.countDocuments({ author: user._id })
     const totalAnswers = await Answer.countDocuments({ author: user._id });
 
     const [questionUpvotes] = await Question.aggregate([
-      { $match: { author: user._id } },
-      {
-        $project: {
-          _id: 0,
-          upvotes: { $size: "$upvotes" },
-        },
-      },
-      {
-        $group: {
-          _id: null,
-          totalUpvotes: { $sum: "$upvotes" },
-        },
-      },
-    ]);
+      { $match: { author: user._id }},
+      { $project: {
+        _id: 0, upvotes: { $size: "$upvotes" }
+      }},
+      { $group: {
+        _id: null,
+        totalUpvotes: { $sum: "$upvotes" }
+      }}
+    ])
 
     const [answerUpvotes] = await Answer.aggregate([
-      { $match: { author: user._id } },
-      {
-        $project: {
-          _id: 0,
-          upvotes: { $size: "$upvotes" },
-        },
-      },
-      {
-        $group: {
-          _id: null,
-          totalUpvotes: { $sum: "$upvotes" },
-        },
-      },
-    ]);
+      { $match: { author: user._id }},
+      { $project: {
+        _id: 0, upvotes: { $size: "$upvotes" }
+      }},
+      { $group: {
+        _id: null,
+        totalUpvotes: { $sum: "$upvotes" }
+      }}
+    ])
 
     const [questionViews] = await Answer.aggregate([
-      { $match: { author: user._id } },
-      {
-        $group: {
-          _id: null,
-          totalViews: { $sum: "$views" },
-        },
-      },
-    ]);
+      { $match: { author: user._id }},
+      { $group: {
+        _id: null,
+        totalViews: { $sum: "$views" }
+      }}
+    ])
 
     const criteria = [
-      { type: "QUESTION_COUNT" as BadgeCriteriaType, count: totalQuestions },
-      { type: "ANSWER_COUNT" as BadgeCriteriaType, count: totalAnswers },
-      {
-        type: "QUESTION_UPVOTES" as BadgeCriteriaType,
-        count: questionUpvotes?.totalUpvotes || 0,
-      },
-      {
-        type: "ANSWER_UPVOTES" as BadgeCriteriaType,
-        count: answerUpvotes?.totalUpvotes || 0,
-      },
-      {
-        type: "TOTAL_VIEWS" as BadgeCriteriaType,
-        count: questionViews?.totalViews || 0,
-      },
-    ];
+      { type: 'QUESTION_COUNT' as BadgeCriteriaType, count: totalQuestions },
+      { type: 'ANSWER_COUNT' as BadgeCriteriaType, count: totalAnswers },
+      { type: 'QUESTION_UPVOTES' as BadgeCriteriaType, count: questionUpvotes?.totalUpvotes || 0 },
+      { type: 'ANSWER_UPVOTES' as BadgeCriteriaType, count: answerUpvotes?.totalUpvotes || 0 },
+      { type: 'TOTAL_VIEWS' as BadgeCriteriaType, count: questionViews?.totalViews || 0 },
+    ]
 
     const badgeCounts = assignBadges({ criteria });
 
@@ -326,7 +298,7 @@ export async function getUserInfo(params: GetUserByIdParams) {
       totalAnswers,
       badgeCounts,
       reputation: user.reputation,
-    };
+    }    
   } catch (error) {
     console.log(error);
     throw error;
@@ -337,19 +309,20 @@ export async function getUserQuestions(params: GetUserStatsParams) {
   try {
     connectToDatabase();
 
-    const { userId, page = 1, pageSize = 2 } = params;
+    const { userId, page = 1, pageSize = 10 } = params;
+
     const skipAmount = (page - 1) * pageSize;
 
-    const totalQuestions = await Question.countDocuments({ author: userId });
+    const totalQuestions = await Question.countDocuments({ author: userId})
 
     const userQuestions = await Question.find({ author: userId })
-      .sort({ createdAt: -1, views: -1, upvotes: -1 })
+      .sort({ createdAt: -1, views: -1, upvotes: -1, })
       .skip(skipAmount)
       .limit(pageSize)
-      .populate("tags", "_id name")
-      .populate("author", "_id clerkId name picture");
+      .populate('tags', '_id name')
+      .populate('author', '_id clerkId name picture')
 
-    const isNextQuestions = totalQuestions > skipAmount + userQuestions.length;
+      const isNextQuestions = totalQuestions > skipAmount + userQuestions.length;
 
     return { totalQuestions, questions: userQuestions, isNextQuestions };
   } catch (error) {
@@ -362,24 +335,33 @@ export async function getUserAnswers(params: GetUserStatsParams) {
   try {
     connectToDatabase();
 
-    const { userId, page = 1, pageSize = 2 } = params;
+    const { userId, page = 1, pageSize = 10 } = params;
 
     const skipAmount = (page - 1) * pageSize;
 
-    const totalAnswers = await Answer.countDocuments({ author: userId });
+    const totalAnswers = await Answer.countDocuments({ author: userId})
 
     const userAnswers = await Answer.find({ author: userId })
       .sort({ upvotes: -1 })
       .skip(skipAmount)
       .limit(pageSize)
-      .populate("question", "_id title")
-      .populate("author", "_id clerkId name picture");
+      .populate('question', '_id title')
+      .populate('author', '_id clerkId name picture')
 
-    const isNextAnswer = totalAnswers > skipAmount + userAnswers.length;
-
+      const isNextAnswer = totalAnswers > skipAmount + userAnswers.length;
+      
     return { totalAnswers, answers: userAnswers, isNextAnswer };
   } catch (error) {
     console.log(error);
     throw error;
   }
 }
+
+// export async function getAllUsers(params: GetAllUsersParams) {
+//   try {
+//     connectToDatabase();
+//   } catch (error) {
+//     console.log(error);
+//     throw error;
+//   }
+// }
